backend/requirements.txt
asgiref==3.11.1
Django==4.2.28
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
